pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "foodnet"
version = "0.1.0"
description = "Swin-Tiny on Food-101 — phases 3+4+5 package"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.2",
    "torchvision>=0.17",
    "timm>=1.0",
    "Pillow>=10.0",
    "numpy>=1.26",
    "pandas>=2.1",
    "pyarrow>=15.0",
    "matplotlib>=3.8",
    "seaborn>=0.13",
    "scikit-learn>=1.4",
    "tqdm>=4.66",
    "wandb>=0.16",
    "torchmetrics>=1.3",
]

[project.optional-dependencies]
dev = ["pytest>=8.0"]

[project.scripts]
foodnet-split = "foodnet.cli.split:main"
foodnet-train = "foodnet.cli.train:main"
foodnet-evaluate = "foodnet.cli.evaluate:main"
foodnet-analyze = "foodnet.cli.analyze:main"

[tool.pytest.ini_options]
markers = ["slow: marks tests as slow"]

[tool.setuptools.packages.find]
include = ["foodnet*"]
